Develop

+++ src/main/java/com/isfive/usearth/domain/maker/service/MakerService.java
@@ -8,6 +8,8 @@ import com.isfive.usearth.domain.maker.entity.Maker;
 import com.isfive.usearth.domain.maker.repository.MakerRepository;
 import com.isfive.usearth.domain.member.entity.Member;
 import com.isfive.usearth.domain.member.repository.MemberRepository;
+import com.isfive.usearth.exception.BusinessException;
+import com.isfive.usearth.exception.ErrorCode;
 import com.isfive.usearth.web.maker.dto.MakerUpdate;
 import com.isfive.usearth.web.maker.dto.register.CorporateRegister;
 import com.isfive.usearth.web.maker.dto.register.IndividualRegister;
@@ -49,15 +51,19 @@ public class MakerService {
 	}
 
 	@Transactional
-	public MakerResponse updateMakerById(Long id, MakerUpdate makerUpdate) {
+	public MakerResponse updateMakerById(String username, Long id, MakerUpdate makerUpdate) {
 		Maker maker = makerRepository.findByIdOrThrow(id);
+		if (!maker.getName().equals(username))
+			throw new BusinessException(ErrorCode.AUTHENTICATION_FAILED);
 		maker.update(makerUpdate);
 		return MakerResponse.fromEntity(makerRepository.save(maker));
 	}
 
 	@Transactional
-	public void removeMakerById(Long id) {
+	public void removeMakerById(String username, Long id) {
 		Maker maker = makerRepository.findByIdOrThrow(id);
+		if (!maker.getName().equals(username))
+			throw new BusinessException(ErrorCode.AUTHENTICATION_FAILED);
 		maker.delete();
 	}
 }
